api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "OpenId connect api implementation"
authors = ["bhavanichandra <bhavanichandra9@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
python-dotenv = "^0.20.0"
requests = "^2.27.1"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
